backend/requirements.txt
flask==3.1.3
flask-cors==6.0.0
Flask-Limiter==3.5.1
gunicorn==23.0.0
isodate==0.7.2
pytz==2025.2
google-api-python-client==2.193.0
requests==2.33.0
firebase-admin==6.6.0
google-cloud-firestore==2.26.0
google-cloud-tasks==2.21.0
pytest==8.4.2
pytest-cov==7.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyJWT==2.12.0
pycryptodome==3.23.0
defusedxml==0.7.1
ruff==0.11.4
